refactor: 汎用コンポーネントを app/components/base/ に移動
Button / Avatar を base/ へ移し、ui/ ・ layout/ と並ぶ3カテゴリに統一。
import パスと story title（Components/Base/...）も更新。
