+++ app/components/base/Avatar/Avatar.test.tsx
