+++ app/components/base/Avatar/index.ts
