+++ app/features/home/pages/HomePage.tsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router'
-import Button from '~/components/Button'
+import Button from '~/components/base/Button'
 
 export default function HomePage() {
   return (
